+++ php/subject/delete.php
@@ -1,5 +1,5 @@
 <?php
-	if (isset($_GET['id'])) {
+	if (isset($_GET['id']) && $_SESSION['logged_role'] == 'manager') {
 		$id = $_GET['id'];
 		require_once '../../php/connection.php';
 		$sql = "DELETE FROM subjects WHERE id='$id'";
@@ -7,5 +7,8 @@
 		$conn->query($sql);
 
 		header('location: '.$_SERVER['HTTP_REFERER']);
+	}
+	else {
+		header('location: ../../views/error/unauthorized.php');
 	}
 ?>
